api_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
aiosqlite==0.20.0
redis==5.0.8
pyjwt==2.9.0
pydantic==2.9.2
onnxruntime==1.18.1
prometheus-client==0.20.0
httpx==0.27.2
pytest==8.3.3
python-multipart==0.0.9
